+++ backend/v1/services/index.ts
@@ -1,0 +1,6 @@
+import { AirPortService } from './airport'
+import { DestinationService } from './destination'
+import { ReservationService } from './reservations'
+import { ResendService } from './email'
+
+export { AirPortService, DestinationService, ReservationService, ResendService }
